style: 방탈출 UI 구현
