+++ src/main/resources/static/admin-reservation.html
@@ -1,0 +1,103 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta charset="UTF-8">
+    <title>Roomescape - 예약 관리</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
+</head>
+<body>
+    <div id="password-overlay" style="position: fixed; top: 0; left: 0; width: 100%; height: 100%; background: white; z-index: 9999; display: flex; align-items: center; justify-content: center;">
+        <div class="card shadow" style="width: 300px;">
+            <div class="card-body">
+                <h5 class="card-title">관리자 인증</h5>
+                <p class="card-text">비밀번호를 입력하세요.</p>
+                <input type="password" id="admin-password" class="form-control mb-3" placeholder="비밀번호" onkeyup="if(event.key === 'Enter') checkPassword()">
+                <button onclick="checkPassword()" class="btn btn-primary w-100">확인</button>
+            </div>
+        </div>
+    </div>
+
+    <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+        <div class="container">
+            <a class="navbar-brand" href="/">Roomescape Admin</a>
+            <div class="navbar-nav">
+                <a class="nav-link" href="/">홈</a>
+                <a class="nav-link" href="/admin-theme.html">테마 관리</a>
+                <a class="nav-link" href="/admin-time.html">시간 관리</a>
+                <a class="nav-link active" href="/admin-reservation.html">예약 관리</a>
+            </div>
+        </div>
+    </nav>
+
+    <div class="container my-5">
+        <h2>예약 관리</h2>
+        
+        <table class="table table-hover mt-4">
+            <thead class="table-dark">
+                <tr>
+                    <th>ID</th>
+                    <th>날짜</th>
+                    <th>시간</th>
+                    <th>테마</th>
+                    <th>예약자</th>
+                    <th>관리</th>
+                </tr>
+            </thead>
+            <tbody id="reservation-table-body">
+                <!-- 예약 목록 -->
+            </tbody>
+        </table>
+    </div>
+
+    <script>
+        function checkPassword() {
+            const password = document.getElementById('admin-password').value;
+            if (password === '1234') {
+                sessionStorage.setItem('admin-authorized', 'true');
+                document.getElementById('password-overlay').style.display = 'none';
+            } else {
+                alert('비밀번호가 틀렸습니다.');
+            }
+        }
+
+        if (sessionStorage.getItem('admin-authorized') === 'true') {
+            document.getElementById('password-overlay').style.display = 'none';
+        }
+
+        document.addEventListener('DOMContentLoaded', fetchReservations);
+
+        function fetchReservations() {
+            fetch('/reservations')
+                .then(response => response.json())
+                .then(reservations => {
+                    const tbody = document.getElementById('reservation-table-body');
+                    tbody.innerHTML = reservations.map(res => `
+                        <tr>
+                            <td>${res.id}</td>
+                            <td>${res.date}</td>
+                            <td>${res.time.startAt}</td>
+                            <td>${res.theme.name}</td>
+                            <td>${res.name}</td>
+                            <td>
+                                <button class="btn btn-danger btn-sm" onclick="deleteReservation(${res.id})">삭제</button>
+                            </td>
+                        </tr>
+                    `).join('');
+                });
+        }
+
+        function deleteReservation(id) {
+            if (!confirm('정말 취소하시겠습니까?')) return;
+            fetch(`/reservations/${id}`, { method: 'DELETE' })
+                .then(response => {
+                    if (response.ok) {
+                        alert('예약이 취소되었습니다.');
+                        fetchReservations();
+                    } else {
+                        alert('예약 취소 실패');
+                    }
+                });
+        }
+    </script>
+</body>
+</html>
