+++ src/main/resources/static/admin-theme.html
@@ -1,0 +1,147 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta charset="UTF-8">
+    <title>Roomescape - 테마 관리</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
+</head>
+<body>
+    <div id="password-overlay" style="position: fixed; top: 0; left: 0; width: 100%; height: 100%; background: white; z-index: 9999; display: flex; align-items: center; justify-content: center;">
+        <div class="card shadow" style="width: 300px;">
+            <div class="card-body">
+                <h5 class="card-title">관리자 인증</h5>
+                <p class="card-text">비밀번호를 입력하세요.</p>
+                <input type="password" id="admin-password" class="form-control mb-3" placeholder="비밀번호" onkeyup="if(event.key === 'Enter') checkPassword()">
+                <button onclick="checkPassword()" class="btn btn-primary w-100">확인</button>
+            </div>
+        </div>
+    </div>
+
+    <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+        <div class="container">
+            <a class="navbar-brand" href="/">Roomescape Admin</a>
+            <div class="navbar-nav">
+                <a class="nav-link" href="/">홈</a>
+                <a class="nav-link active" href="/admin-theme.html">테마 관리</a>
+                <a class="nav-link" href="/admin-time.html">시간 관리</a>
+                <a class="nav-link" href="/admin-reservation.html">예약 관리</a>
+            </div>
+        </div>
+    </nav>
+
+    <div class="container my-5">
+        <h2>테마 관리</h2>
+        
+        <div class="card my-4 shadow-sm">
+            <div class="card-body">
+                <h5 class="card-title">새 테마 추가</h5>
+                <form id="theme-form">
+                    <div class="mb-3">
+                        <label class="form-label">테마 이름</label>
+                        <input type="text" id="theme-name" class="form-control" required>
+                    </div>
+                    <div class="mb-3">
+                        <label class="form-label">설명</label>
+                        <textarea id="theme-description" class="form-control" required></textarea>
+                    </div>
+                    <div class="mb-3">
+                        <label class="form-label">썸네일 URL</label>
+                        <input type="url" id="theme-thumbnail" class="form-control" required>
+                    </div>
+                    <button type="submit" class="btn btn-primary">추가하기</button>
+                </form>
+            </div>
+        </div>
+
+        <table class="table table-hover mt-4">
+            <thead class="table-dark">
+                <tr>
+                    <th>ID</th>
+                    <th>썸네일</th>
+                    <th>이름</th>
+                    <th>설명</th>
+                    <th>관리</th>
+                </tr>
+            </thead>
+            <tbody id="theme-table-body">
+                <!-- 테마 목록 -->
+            </tbody>
+        </table>
+    </div>
+
+    <script>
+        function checkPassword() {
+            const password = document.getElementById('admin-password').value;
+            if (password === '1234') {
+                sessionStorage.setItem('admin-authorized', 'true');
+                document.getElementById('password-overlay').style.display = 'none';
+            } else {
+                alert('비밀번호가 틀렸습니다.');
+            }
+        }
+
+        if (sessionStorage.getItem('admin-authorized') === 'true') {
+            document.getElementById('password-overlay').style.display = 'none';
+        }
+
+        document.addEventListener('DOMContentLoaded', fetchThemes);
+        document.getElementById('theme-form').addEventListener('submit', addTheme);
+
+        function fetchThemes() {
+            fetch('/themes')
+                .then(response => response.json())
+                .then(themes => {
+                    const tbody = document.getElementById('theme-table-body');
+                    tbody.innerHTML = themes.map(theme => `
+                        <tr>
+                            <td>${theme.id}</td>
+                            <td><img src="${theme.thumbnail}" width="50" height="50" style="object-fit: cover;" onerror="this.src='https://via.placeholder.com/50?text=No'"></td>
+                            <td>${theme.name}</td>
+                            <td>${theme.description}</td>
+                            <td>
+                                <button class="btn btn-danger btn-sm" onclick="deleteTheme(${theme.id})">삭제</button>
+                            </td>
+                        </tr>
+                    `).join('');
+                });
+        }
+
+        function addTheme(event) {
+            event.preventDefault();
+            const payload = {
+                name: document.getElementById('theme-name').value,
+                description: document.getElementById('theme-description').value,
+                thumbnail: document.getElementById('theme-thumbnail').value
+            };
+
+            fetch('/admin/themes', {
+                method: 'POST',
+                headers: { 'Content-Type': 'application/json' },
+                body: JSON.stringify(payload)
+            })
+            .then(response => {
+                if (response.ok) {
+                    alert('테마가 추가되었습니다.');
+                    document.getElementById('theme-form').reset();
+                    fetchThemes();
+                } else {
+                    alert('테마 추가 실패');
+                }
+            });
+        }
+
+        function deleteTheme(id) {
+            if (!confirm('정말 삭제하시겠습니까?')) return;
+            fetch(`/admin/themes/${id}`, { method: 'DELETE' })
+                .then(response => {
+                    if (response.ok) {
+                        alert('테마가 삭제되었습니다.');
+                        fetchThemes();
+                    } else {
+                        alert('테마 삭제 실패');
+                    }
+                });
+        }
+    </script>
+</body>
+</html>
