+++ src/main/resources/static/admin-time.html
@@ -1,0 +1,165 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta charset="UTF-8">
+    <title>Roomescape - 시간 관리</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
+</head>
+<body>
+    <div id="password-overlay" style="position: fixed; top: 0; left: 0; width: 100%; height: 100%; background: white; z-index: 9999; display: flex; align-items: center; justify-content: center;">
+        <div class="card shadow" style="width: 300px;">
+            <div class="card-body">
+                <h5 class="card-title">관리자 인증</h5>
+                <p class="card-text">비밀번호를 입력하세요.</p>
+                <input type="password" id="admin-password" class="form-control mb-3" placeholder="비밀번호" onkeyup="if(event.key === 'Enter') checkPassword()">
+                <button onclick="checkPassword()" class="btn btn-primary w-100">확인</button>
+            </div>
+        </div>
+    </div>
+
+    <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+        <div class="container">
+            <a class="navbar-brand" href="/">Roomescape Admin</a>
+            <div class="navbar-nav">
+                <a class="nav-link" href="/">홈</a>
+                <a class="nav-link" href="/admin-theme.html">테마 관리</a>
+                <a class="nav-link active" href="/admin-time.html">시간 관리</a>
+                <a class="nav-link" href="/admin-reservation.html">예약 관리</a>
+            </div>
+        </div>
+    </nav>
+
+    <div class="container my-5">
+        <h2>시간 관리</h2>
+        
+        <div class="card my-4 shadow-sm">
+            <div class="card-body">
+                <h5 class="card-title">새 시간 추가</h5>
+                <form id="time-form" class="row g-3">
+                    <div class="col-auto">
+                        <select id="time-start" class="form-select" required>
+                            <option value="" disabled selected>시간 선택</option>
+                            <option value="10:00">10:00</option>
+                            <option value="11:00">11:00</option>
+                            <option value="12:00">12:00</option>
+                            <option value="13:00">13:00</option>
+                            <option value="14:00">14:00</option>
+                            <option value="15:00">15:00</option>
+                            <option value="16:00">16:00</option>
+                            <option value="17:00">17:00</option>
+                            <option value="18:00">18:00</option>
+                        </select>
+                    </div>
+                    <div class="col-auto">
+                        <button type="submit" class="btn btn-primary">추가하기</button>
+                    </div>
+                </form>
+            </div>
+        </div>
+
+        <table class="table table-hover mt-4">
+            <thead class="table-dark">
+                <tr>
+                    <th>ID</th>
+                    <th>시작 시간</th>
+                    <th>관리</th>
+                </tr>
+            </thead>
+            <tbody id="time-table-body">
+                <!-- 시간 목록 -->
+            </tbody>
+        </table>
+    </div>
+
+    <script>
+        function checkPassword() {
+            const password = document.getElementById('admin-password').value;
+            if (password === '1234') {
+                sessionStorage.setItem('admin-authorized', 'true');
+                document.getElementById('password-overlay').style.display = 'none';
+            } else {
+                alert('비밀번호가 틀렸습니다.');
+            }
+        }
+
+        if (sessionStorage.getItem('admin-authorized') === 'true') {
+            document.getElementById('password-overlay').style.display = 'none';
+        }
+
+        document.addEventListener('DOMContentLoaded', fetchAllTimes);
+        document.getElementById('time-form').addEventListener('submit', addTime);
+
+        function fetchAllTimes() {
+            // 모든 시간을 가져오기 위해 첫 번째 테마를 먼저 조회함
+            fetch('/themes')
+                .then(response => response.json())
+                .then(themes => {
+                    if (themes.length === 0) {
+                        alert('테마를 먼저 등록해주세요.');
+                        return;
+                    }
+                    const themeId = themes[0].id;
+                    const date = '2099-12-31'; // 예약이 없을 법한 미래 날짜
+                    return fetch(`/times?themeId=${themeId}&date=${date}`);
+                })
+                .then(response => response.json())
+                .then(times => {
+                    if (!times) return;
+                    const tbody = document.getElementById('time-table-body');
+                    tbody.innerHTML = times.map(time => `
+                        <tr>
+                            <td>${time.id || time.timeId}</td>
+                            <td>${time.startAt}</td>
+                            <td>
+                                <button class="btn btn-danger btn-sm" onclick="deleteTime(${time.id || time.timeId})">삭제</button>
+                            </td>
+                        </tr>
+                    `).join('');
+                });
+        }
+
+        function addTime(event) {
+            event.preventDefault();
+            const timeValue = document.getElementById('time-start').value;
+            const hour = parseInt(timeValue.split(':')[0]);
+            const minute = parseInt(timeValue.split(':')[1]);
+
+            if (hour < 10 || hour > 18 || minute !== 0) {
+                alert('10시부터 18시까지 1시간 단위로만 예약할 수 있습니다. (예: 10:00, 11:00, ..., 18:00)');
+                return;
+            }
+
+            const payload = {
+                startAt: timeValue
+            };
+
+            fetch('/admin/times', {
+                method: 'POST',
+                headers: { 'Content-Type': 'application/json' },
+                body: JSON.stringify(payload)
+            })
+            .then(response => {
+                if (response.ok) {
+                    alert('시간이 추가되었습니다.');
+                    fetchAllTimes();
+                } else {
+                    alert('시간 추가 실패');
+                }
+            });
+        }
+
+        function deleteTime(id) {
+            if (!confirm('정말 삭제하시겠습니까?')) return;
+            fetch(`/admin/times/${id}`, { method: 'DELETE' })
+                .then(response => {
+                    if (response.ok) {
+                        alert('시간이 삭제되었습니다.');
+                        fetchAllTimes();
+                    } else {
+                        alert('시간 삭제 실패');
+                    }
+                });
+        }
+    </script>
+</body>
+</html>
