+++ src/main/resources/static/reservation.html
@@ -1,0 +1,170 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta charset="UTF-8">
+    <title>Roomescape - 예약하기</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
+    <style>
+        body { background-color: #f8f9fa; }
+        .reservation-container { max-width: 800px; margin: 3rem auto; padding: 2rem; background: white; border-radius: 10px; shadow: 0 0.5rem 1rem rgba(0,0,0,0.15); }
+        .theme-selected { border: 2px solid #0d6efd !important; }
+        .time-slot { cursor: pointer; transition: all 0.2s; }
+        .time-slot:not(.disabled):hover { background-color: #e9ecef; }
+        .time-slot.selected { background-color: #0d6efd; color: white; border-color: #0d6efd; }
+    </style>
+</head>
+<body>
+    <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+        <div class="container">
+            <a class="navbar-brand" href="/">Roomescape</a>
+            <div class="navbar-nav">
+                <a class="nav-link" href="/">홈</a>
+                <a class="nav-link active" href="/reservation.html">예약하기</a>
+            </div>
+        </div>
+    </nav>
+
+    <div class="container">
+        <div class="reservation-container shadow">
+            <h2 class="text-center mb-4">방탈출 예약</h2>
+            
+            <div class="mb-4">
+                <h5>1. 테마 선택</h5>
+                <div id="theme-list" class="row g-3 mt-1">
+                    <!-- 테마 목록 -->
+                </div>
+            </div>
+
+            <div class="mb-4">
+                <h5>2. 날짜 선택</h5>
+                <input type="date" id="reservation-date" class="form-control" min="">
+            </div>
+
+            <div class="mb-4">
+                <h5>3. 시간 선택</h5>
+                <div id="time-list" class="row g-2 mt-1">
+                    <p class="text-muted">테마와 날짜를 먼저 선택해주세요.</p>
+                </div>
+            </div>
+
+            <div class="mb-4">
+                <h5>4. 예약자 정보</h5>
+                <input type="text" id="reservation-name" class="form-control" placeholder="이름을 입력하세요">
+            </div>
+
+            <button id="submit-reservation" class="btn btn-primary w-100 btn-lg" disabled>예약 완료</button>
+        </div>
+    </div>
+
+    <script>
+        let selectedThemeId = null;
+        let selectedTimeId = null;
+
+        document.addEventListener('DOMContentLoaded', () => {
+            const dateInput = document.getElementById('reservation-date');
+            const today = new Date().toISOString().split('T')[0];
+            dateInput.min = today;
+            dateInput.value = today;
+
+            fetchThemes();
+
+            dateInput.addEventListener('change', fetchTimes);
+            document.getElementById('submit-reservation').addEventListener('click', submitReservation);
+        });
+
+        function fetchThemes() {
+            fetch('/themes')
+                .then(response => response.json())
+                .then(themes => {
+                    const themeList = document.getElementById('theme-list');
+                    themeList.innerHTML = themes.map(theme => `
+                        <div class="col-md-4">
+                            <div class="card h-100 theme-item" data-id="${theme.id}" style="cursor: pointer;">
+                                <img src="${theme.thumbnail}" class="card-img-top" alt="${theme.name}" height="100" style="object-fit: cover;" onerror="this.src='https://via.placeholder.com/150?text=No+Image'">
+                                <div class="card-body p-2 text-center">
+                                    <h6 class="card-title mb-0">${theme.name}</h6>
+                                </div>
+                            </div>
+                        </div>
+                    `).join('');
+
+                    document.querySelectorAll('.theme-item').forEach(item => {
+                        item.addEventListener('click', function() {
+                            document.querySelectorAll('.theme-item').forEach(i => i.classList.remove('theme-selected'));
+                            this.classList.add('theme-selected');
+                            selectedThemeId = this.dataset.id;
+                            fetchTimes();
+                        });
+                    });
+                });
+        }
+
+        function fetchTimes() {
+            const date = document.getElementById('reservation-date').value;
+            if (!selectedThemeId || !date) return;
+
+            fetch(`/times?themeId=${selectedThemeId}&date=${date}`)
+                .then(response => response.json())
+                .then(times => {
+                    const timeList = document.getElementById('time-list');
+                    if (times.length === 0) {
+                        timeList.innerHTML = '<p class="text-danger">해당 날짜에 가능한 시간이 없습니다.</p>';
+                        return;
+                    }
+                    timeList.innerHTML = times.map(time => `
+                        <div class="col-4">
+                            <div class="border rounded p-2 text-center time-slot ${time.isNotReserved ? '' : 'disabled bg-light text-muted'}" 
+                                 data-id="${time.id}" 
+                                 ${!time.isNotReserved ? 'style="pointer-events: none;"' : ''}>
+                                ${time.startAt}
+                            </div>
+                        </div>
+                    `).join('');
+
+                    document.querySelectorAll('.time-slot:not(.disabled)').forEach(slot => {
+                        slot.addEventListener('click', function() {
+                            document.querySelectorAll('.time-slot').forEach(s => s.classList.remove('selected'));
+                            this.classList.add('selected');
+                            selectedTimeId = this.dataset.id;
+                            checkValidation();
+                        });
+                    });
+                });
+        }
+
+        function checkValidation() {
+            const name = document.getElementById('reservation-name').value;
+            document.getElementById('submit-reservation').disabled = !(selectedThemeId && selectedTimeId && name);
+        }
+
+        document.getElementById('reservation-name').addEventListener('input', checkValidation);
+
+        function submitReservation() {
+            const name = document.getElementById('reservation-name').value;
+            const date = document.getElementById('reservation-date').value;
+
+            const payload = {
+                name: name,
+                date: date,
+                timeId: parseInt(selectedTimeId),
+                themeId: parseInt(selectedThemeId)
+            };
+
+            fetch('/reservations', {
+                method: 'POST',
+                headers: { 'Content-Type': 'application/json' },
+                body: JSON.stringify(payload)
+            })
+            .then(response => {
+                if (response.ok) {
+                    alert('예약이 완료되었습니다!');
+                    location.href = '/';
+                } else {
+                    return response.json().then(err => { throw new Error(err.message || '예약 실패'); });
+                }
+            })
+            .catch(error => alert(error.message));
+        }
+    </script>
+</body>
+</html>
