+++ src/test/java/roomescape/controller/AdminReservationTimeControllerTest.java
@@ -1,0 +1,49 @@
+package roomescape.controller;
+
+import static org.hamcrest.Matchers.notNullValue;
+
+import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import java.util.HashMap;
+import java.util.Map;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class AdminReservationTimeControllerTest {
+
+    @Test
+    void 시간_추가() {
+        Map<String, String> params = new HashMap<>();
+        params.put("startAt", "10:00");
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/admin/times")
+                .then().log().all()
+                .statusCode(201)
+                .body("id", notNullValue());
+    }
+
+    @Test
+    void 시간_삭제() {
+        Map<String, String> params = new HashMap<>();
+        params.put("startAt", "10:00");
+
+        int timeId = RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/admin/times")
+                .then().log().all()
+                .statusCode(201)
+                .extract().path("id");
+
+        RestAssured.given().log().all()
+                .when().delete("/admin/times/" + timeId)
+                .then().log().all()
+                .statusCode(204);
+    }
+}
